added zip packaging to deploy script

+++ deploy.sh
@@ -1,6 +1,13 @@
 #!/bin/sh
 
-DEST=~/.local/share/gnome-shell/extensions/keyman@dpoetzsch.github.com
+MODE="deploy"
+if [ $1 = "zip" ]; then
+    MODE="zip"
+fi
+
+FILES="extension.js keyman.js clipboard.js keyringDbus.js keyringInterfaces.js"
+FILES="$FILES utils.js data.js settings.js prefs.js metadata.json"
+FILES="$FILES LICENSE README.md schemas/ locale/ keyman.pot stylesheet.css"
 
 # compile locales
 echo "Compiling locales..."
@@ -9,16 +16,22 @@ for locpath in locale/*; do
     msgfmt -o $locpath/LC_MESSAGES/keyman.mo $locpath/LC_MESSAGES/keyman.po
 done
 
+echo "Compiling schemas..."
 glib-compile-schemas schemas/
 
-echo "Removing old instance..."
-rm -r $DEST
-mkdir $DEST
+if [ $MODE = "deploy" ]; then
+    DEST=~/.local/share/gnome-shell/extensions/keyman@dpoetzsch.github.com
 
-echo "Copying content..."
-cp -a extension.js keyman.js clipboard.js keyringDbus.js keyringInterfaces.js \
-      utils.js data.js settings.js prefs.js metadata.json \
-      LICENSE README.md schemas/ locale/ keyman.pot stylesheet.css $DEST/
-       
-echo "Restarting gnome shell..."
-gnome-shell --replace &
+    echo "Removing old instance..."
+    rm -r $DEST
+    mkdir $DEST
+
+    echo "Copying content..."
+    cp -a $FILES $DEST/
+           
+    echo "Restarting gnome shell..."
+    gnome-shell --replace &
+else
+    echo "Creating keyman.zip..."
+    zip -r keyman.zip $FILES
+fi
